requirements.txt
Pillow
pillow_heif
